206-ninja_query_in/pyproject.toml
[tool.poetry]
name = "205-ninja_query_in"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "205_ninja_query_in"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2"
django-ninja = "^0.22.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
